MA-20704 fix ethernet build error on android13
/* This is a temporary version */
EthernetManager is going to be moved to Connectivity mainline module, some api are
unsupported on android13, we need to use the new exposed ones to replace the old ones.
1. use addInterfaceStateListener to replace addListener.
2. rely on the onInterfaceStateChanged callback to receive the Ethernet interface state
   update, to replace the getAvailableInterfaces and getConfiguration.
3. use updateConfiguration to replace setConfiguration.

this patch can get ethernet config, but if we manually config the ethernet from the dialog,
fail to update the IpConfig, due to updateConfiguration seems only work on an automotive
device, I will try to use another way to fix it later.

Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Change-Id: I7c1ed63fae995c1e53f19d2c6f57924b5f83c8c7
Reviewed-on: http://androidsource.ap.freescale.net/project/20004
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: <yunjie.jia@nxp.com>